+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - chore: Change IpfsOptions visibility, remove UninitializedIpfs::{empty, with_opt}. [PR 294](https://github.com/dariusc93/rust-ipfs/pull/294)
 - refactor: Support multiple certificates for wss. [PR 295](https://github.com/dariusc93/rust-ipfs/pull/295)
 - refactor: Remove redundant static lifetime. [PR 301](https://github.com/dariusc93/rust-ipfs/pull/301)
+- refactor: Remove optional error from `UnixfsStatus`.
 
 # 0.11.21
 - chore: Put libp2p-webrtc-websys behind feature.